backend/requirements.txt
# MailLens Python dependencies

# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
psycopg[binary]>=3.2.0
pgvector>=0.3.0

# LLM providers
anthropic>=0.40.0
openai>=1.50.0
google-genai>=1.0.0

# HTTP client (for Ollama)
httpx>=0.27.0

# Config
pyyaml>=6.0
pydantic>=2.0

# Email parsing
email-reply-parser>=0.5.0

# PDF text extraction
pymupdf>=1.24.0

# Utilities
python-multipart>=0.0.9
